requirements.txt
cocos2d>=0.4.0
networkx>=1.3
pyglet>=1.1.4
